Move platform API to its own directory
The serializer classes can probably be moved up since they are generic and could
be widely used.
